Update chrome-driver path, last commit using Selenium
I updated the chrome-driver path because I needed to shift the location of the chromium
folder back to /usr/lib/ but the Rapsberry Pi version required Chromium version 84. I will
stop using Selenium and will be shifting over to beautifulsoup and requests for this
project. Next commit will replace most of the Selenium code.